# api/openapi.yaml
openapi: 3.1.0
info:
  title: Origami FarmOS API
  description: Offline-first, tablet-first farm operating system API for Origami Farms.
    See CONSTITUTION.md and product/MVP_SCOPE.md in the repository root for the product
    principles this API enforces (evidence-based recommendations, worker-observes/manager-decides
    separation, event-sourced history).
  version: 0.1.0
paths:
...
  /api/v1/fields:
    post:
      tags:
      - agriculture
      summary: Create Field
      operationId: create_field_api_v1_fields_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/fields/{field_id}:
    patch:
      tags:
      - agriculture
      summary: Update Field
      operationId: update_field_api_v1_fields__field_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: field_id
        in: path
        required: true
        schema:
          type: string
          title: Field Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/crops:
    get:
      tags:
      - agriculture
      summary: List Crops
      operationId: list_crops_api_v1_crops_get
      security:
      - HTTPBearer: []
      parameters:
      - name: include_inactive
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Inactive
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CropOut'
                title: Response List Crops Api V1 Crops Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - agriculture
      summary: Create Crop
      operationId: create_crop_api_v1_crops_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CropCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CropOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/crops/{crop_id}:
    delete:
      tags:
      - agriculture
      summary: Archive Crop
      operationId: archive_crop_api_v1_crops__crop_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: crop_id
        in: path
        required: true
        schema:
          type: string
          title: Crop Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/crop-plantings:
    get:
      tags:
      - agriculture
      summary: List Plantings
      operationId: list_plantings_api_v1_crop_plantings_get
      security:
      - HTTPBearer: []
      parameters:
      - name: field_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Id
      - name: active_only
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Active Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CropPlantingOut'
                title: Response List Plantings Api V1 Crop Plantings Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - agriculture
      summary: Create Planting
      operationId: create_planting_api_v1_crop_plantings_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CropPlantingCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CropPlantingOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/crop-plantings/{planting_id}:
    patch:
      tags:
      - agriculture
      summary: Update Planting
      operationId: update_planting_api_v1_crop_plantings__planting_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: planting_id
        in: path
        required: true
        schema:
          type: string
          title: Planting Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CropPlantingUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CropPlantingOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/harvest:
    post:
      tags:
      - agriculture
      summary: Record Daily Harvest
      description: 'Records the day''s pick and moves the sellable part into real

        inventory, so "168 kg ready for sale" on the Produce screen is stock

        that exists rather than a number someone typed.'
      operationId: record_daily_harvest_api_v1_harvest_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyHarvestCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyHarvestOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    CropCreate:
      properties:
        name:
          type: string
          title: Name
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        default_cycle_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Default Cycle Days
      type: object
      required:
      - name
      title: CropCreate
      description: "Tech spec \xA716: crop types are farm data, never a hard-coded\
        \ list."
    CropOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        default_cycle_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Default Cycle Days
        active:
          type: boolean
          title: Active
      type: object
      required:
      - id
      - name
      - active
      title: CropOut
    CropPlantingCreate:
      properties:
        field_id:
          type: string
          title: Field Id
        crop_id:
          type: string
          title: Crop Id
        variety:
          anyOf:
          - type: string
          - type: 'null'
          title: Variety
        planted_area:
          anyOf:
          - type: number
          - type: 'null'
          title: Planted Area
        area_unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Area Unit
          default: m2
        planted_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Planted Date
        expected_harvest_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expected Harvest Date
        expected_yield_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Expected Yield Kg
        stage:
          type: string
          title: Stage
          default: planted
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - field_id
      - crop_id
      title: CropPlantingCreate
    CropPlantingOut:
      properties:
        id:
          type: string
          title: Id
        field_id:
          type: string
          title: Field Id
        crop_id:
          type: string
          title: Crop Id
        variety:
          anyOf:
          - type: string
          - type: 'null'
          title: Variety
        planted_area:
          anyOf:
          - type: number
          - type: 'null'
          title: Planted Area
        area_unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Area Unit
        planted_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Planted Date
        expected_harvest_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expected Harvest Date
        expected_yield_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Expected Yield Kg
        stage:
          type: string
          title: Stage
        status:
          type: string
          title: Status
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - field_id
      - crop_id
      - stage
      - status
      - created_at
      title: CropPlantingOut
    CropPlantingUpdate:
      properties:
        variety:
          anyOf:
          - type: string
          - type: 'null'
          title: Variety
        planted_area:
          anyOf:
          - type: number
          - type: 'null'
          title: Planted Area
        expected_harvest_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expected Harvest Date
        expected_yield_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Expected Yield Kg
        stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Stage
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      title: CropPlantingUpdate
...
    DailyHarvestCreate:
      properties:
        field_id:
          type: string
          title: Field Id
        planting_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Planting Id
        crop_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Crop Id
        product_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Name
        total_quantity:
          type: number
          title: Total Quantity
        sellable_quantity:
          anyOf:
          - type: number
          - type: 'null'
          title: Sellable Quantity
        waste_quantity:
          type: number
          title: Waste Quantity
          default: 0
        unit:
          type: string
          title: Unit
          default: kg
        destination:
          anyOf:
          - type: string
          - type: 'null'
          title: Destination
        recorded_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Recorded At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - field_id
      - total_quantity
      title: DailyHarvestCreate
      description: "Tech spec \xA717 \"How much was harvested today?\" \u2014 the\
        \ sellable/waste\nsplit is the point: it is what makes waste and cost-per-kg\
        \ real\nnumbers instead of estimates."
    DailyHarvestOut:
      properties:
        id:
          type: string
          title: Id
        field_id:
          type: string
          title: Field Id
        product_name:
          type: string
          title: Product Name
        total_quantity:
          type: number
          title: Total Quantity
        sellable_quantity:
          type: number
          title: Sellable Quantity
        waste_quantity:
          type: number
          title: Waste Quantity
        unit:
          type: string
          title: Unit
        recorded_at:
          type: string
          format: date-time
          title: Recorded At
        inventory_item_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Inventory Item Id
        inventory_qty_after:
          anyOf:
          - type: number
          - type: 'null'
          title: Inventory Qty After
      type: object
      required:
      - id
      - field_id
      - product_name
      - total_quantity
      - sellable_quantity
      - waste_quantity
      - unit
      - recorded_at
      title: DailyHarvestOut
...
    FieldCreate:
      properties:
        name:
          type: string
          title: Name
        field_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Code
        area_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Area Value
        area_unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Area Unit
          default: m2
        location_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Label
        soil_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Soil Type
        irrigation_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Irrigation Method
        status:
          type: string
          title: Status
          default: active
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - name
      title: FieldCreate
    FieldOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        crop_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Crop Type
        area_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Area Value
        area_unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Area Unit
        stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Stage
        expected_harvest_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expected Harvest Date
        est_yield_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Est Yield Kg
        field_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Code
        location_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Label
        soil_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Soil Type
        irrigation_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Irrigation Method
        status:
          type: string
          title: Status
          default: active
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - id
      - name
      title: FieldOut
    FieldUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        field_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Code
        area_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Area Value
        area_unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Area Unit
        location_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Label
        soil_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Soil Type
        irrigation_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Irrigation Method
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        crop_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Crop Type
        stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Stage
        expected_harvest_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expected Harvest Date
        est_yield_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Est Yield Kg
      type: object
      title: FieldUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer